refactor code, use all features to calculate level for compats by make
